feat: add schema
